ref(kv_parser): rename file to be more explicit
