+++ src/transforms/mod.rs
@@ -30,7 +30,7 @@ pub mod grok_parser;
 #[cfg(feature = "transforms-json_parser")]
 pub mod json_parser;
 #[cfg(feature = "transforms-kv_parser")]
-pub mod kv_parser;
+pub mod key_value_parser;
 #[cfg(feature = "transforms-log_to_metric")]
 pub mod log_to_metric;
 #[cfg(feature = "transforms-logfmt_parser")]
